Change the way of getting token file in the task and remove VisitedClasses related functionality

+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/TaskUpdateImports.java
@@ -49,7 +49,6 @@ import elemental.json.JsonObject;
 import elemental.json.impl.JsonUtil;
 
 import static com.vaadin.flow.server.frontend.FrontendUtils.IMPORTS_NAME;
-import static com.vaadin.flow.server.frontend.FrontendUtils.TOKEN_FILE;
 
 /**
  * An updater that it's run when the servlet context is initialised in dev-mode
@@ -73,7 +72,7 @@ public class TaskUpdateImports extends NodeUpdater {
     private final File frontendDirectory;
     private final FrontendDependenciesScanner fallbackScanner;
     private final ClassFinder finder;
-    private final File webpackOutputDirectory;
+    private final File tokenFile;
     private final JsonObject tokenFileData;
 
     private class UpdateMainImportsFile extends AbstractUpdateImports {
@@ -278,16 +277,16 @@ public class TaskUpdateImports extends NodeUpdater {
      *            folder where flow generated files will be placed.
      * @param frontendDirectory
      *            a directory with project's frontend files
-     * @param webpackOutputDirectory
-     *            the directory to set for webpack to output its build results.
+     * @param tokenFile
+     *            the token (flow-build-info.json) path, may be {@code null}
      */
     TaskUpdateImports(ClassFinder finder,
             FrontendDependenciesScanner frontendDepScanner,
             SerializableFunction<ClassFinder, FrontendDependenciesScanner> fallBackScannerProvider,
             File npmFolder, File generatedPath, File frontendDirectory,
-            File webpackOutputDirectory) {
+            File tokenFile) {
         this(finder, frontendDepScanner, fallBackScannerProvider, npmFolder,
-                generatedPath, frontendDirectory, webpackOutputDirectory, null);
+                generatedPath, frontendDirectory, tokenFile, null);
     }
 
     /**
@@ -305,8 +304,8 @@ public class TaskUpdateImports extends NodeUpdater {
      *            folder where flow generated files will be placed.
      * @param frontendDirectory
      *            a directory with project's frontend files
-     * @param webpackOutputDirectory
-     *            the directory to set for webpack to output its build results.
+     * @param tokenFile
+     *            the token (flow-build-info.json) path, may be {@code null}
      * @param tokenFileData
      *            object to fill with token file data, may be {@code null}
      */
@@ -314,12 +313,12 @@ public class TaskUpdateImports extends NodeUpdater {
             FrontendDependenciesScanner frontendDepScanner,
             SerializableFunction<ClassFinder, FrontendDependenciesScanner> fallBackScannerProvider,
             File npmFolder, File generatedPath, File frontendDirectory,
-            File webpackOutputDirectory, JsonObject tokenFileData) {
+            File tokenFile, JsonObject tokenFileData) {
         super(finder, frontendDepScanner, npmFolder, generatedPath);
         this.frontendDirectory = frontendDirectory;
         fallbackScanner = fallBackScannerProvider.apply(finder);
         this.finder = finder;
-        this.webpackOutputDirectory = webpackOutputDirectory;
+        this.tokenFile = tokenFile;
         this.tokenFileData = tokenFileData;
     }
 
@@ -368,30 +367,26 @@ public class TaskUpdateImports extends NodeUpdater {
     }
 
     private void updateBuildFile(AbstractUpdateImports updater) {
-        File tokenFile = getTokenFile();
-        if (!tokenFile.exists()) {
-            log().warn("Missing token file. New token file will be created.");
+        boolean tokenFileExists = tokenFile != null && tokenFile.exists();
+        if (!tokenFileExists) {
+            log().warn(
+                    "Token file is not available. Fallback chunk data won't be written.");
         }
         try {
-            JsonObject buildInfo;
-            if (tokenFile.exists()) {
+            if (tokenFileExists) {
                 String json = FileUtils.readFileToString(tokenFile,
                         StandardCharsets.UTF_8);
-                buildInfo = JsonUtil.parse(json);
-            } else {
-                FileUtils.forceMkdirParent(tokenFile);
-                buildInfo = Json.createObject();
+                JsonObject buildInfo = JsonUtil.parse(json);
+                populateFallbackData(buildInfo, updater);
+                FileUtils.write(tokenFile, JsonUtil.stringify(buildInfo, 2),
+                        StandardCharsets.UTF_8);
             }
 
-            populateFallbackData(buildInfo, updater);
-            if (tokenFileData != null) {
-                populateFallbackData(tokenFileData, updater);
-            }
-
-            FileUtils.write(tokenFile, JsonUtil.stringify(buildInfo, 2),
-                    StandardCharsets.UTF_8);
         } catch (IOException e) {
             log().warn("Unable to read token file", e);
+        }
+        if (tokenFileData != null) {
+            populateFallbackData(tokenFileData, updater);
         }
     }
 
@@ -453,7 +448,4 @@ public class TaskUpdateImports extends NodeUpdater {
         return object;
     }
 
-    private File getTokenFile() {
-        return new File(webpackOutputDirectory, TOKEN_FILE);
-    }
 }
